requirements.txt
openai==1.107.2
python-dotenv==1.1.1
Requests==2.32.5
streamlit==1.49.1
